........
r2556 | marco | 2006-02-12 23:00:38 +0100 (Sun, 12 Feb 2006) | 2 lines

 * dbf_wtil reenabled for x86_64

........
r2557 | peter | 2006-02-13 08:19:07 +0100 (Mon, 13 Feb 2006) | 2 lines

  * remove duplicate dbf_wtil

........
r2558 | peter | 2006-02-13 10:15:11 +0100 (Mon, 13 Feb 2006) | 3 lines

  * make solaris zipinstall working
  * remove ezcgi.inc, use GetEnvironmentVariable from SysUtils

........
r2604 | joost | 2006-02-16 00:24:09 +0100 (Thu, 16 Feb 2006) | 1 line

 + implemented TOracleConnection
........
r2630 | joost | 2006-02-19 16:08:00 +0100 (Sun, 19 Feb 2006) | 1 line

 + Do not handle :: in a query as a parameter (bug 4813-2)
........
r2631 | joost | 2006-02-19 16:22:21 +0100 (Sun, 19 Feb 2006) | 3 lines

 + added empty TSQLConnection.GetHandle
 + added empty TSQLConnection.CreateBlobStream
 + implemented TSQLConnection.StartTransaction and EndTransaction
........
r2635 | joost | 2006-02-19 18:42:26 +0100 (Sun, 19 Feb 2006) | 1 line

+ fixed an enabled oracle for win32
........
r2638 | florian | 2006-02-19 22:33:59 +0100 (Sun, 19 Feb 2006) | 2 lines

* ar patch from Yury

........
r2640 | peter | 2006-02-20 07:59:14 +0100 (Mon, 20 Feb 2006) | 2 lines

  * fix range error

........
r2649 | jonas | 2006-02-20 11:39:37 +0100 (Mon, 20 Feb 2006) | 2 lines

  * fixed range errors

........
r2657 | joost | 2006-02-20 17:53:35 +0100 (Mon, 20 Feb 2006) | 1 line

+ Fixed mysql_error
........
r2660 | joost | 2006-02-21 11:14:31 +0100 (Tue, 21 Feb 2006) | 1 line

 + Moved obsolete DB-units to unmaintained
........
r2661 | joost | 2006-02-21 11:17:47 +0100 (Tue, 21 Feb 2006) | 1 line

 + renamed tests to ddg
........
r2662 | joost | 2006-02-21 15:29:30 +0100 (Tue, 21 Feb 2006) | 1 line

 * Fix for error-handling when a TSQLQuery is opened.
........
r2663 | joost | 2006-02-21 22:17:11 +0100 (Tue, 21 Feb 2006) | 4 lines

 + made TDataset.CreateFields virtual (delphi compat)
 + implemented TParam.AsLargeInt
 + cleanup of IBConnection LargeInt code
 + implemented LargeInt parameter type for IBConnection
........
r2664 | joost | 2006-02-22 12:58:35 +0100 (Wed, 22 Feb 2006) | 1 line

 + forgot to include changes in db.pp and dsparams.inc in r2663
........
r2678 | daniel | 2006-02-25 00:31:55 +0100 (Sat, 25 Feb 2006) | 2 lines

  * Optimize keyboard unit for code size; now 3kb+ object code less.

........
r2706 | hajny | 2006-02-28 00:43:30 +0100 (Tue, 28 Feb 2006) | 1 line

  * attempt to fix broken compilation for go32v2 target
........
r2719 | daniel | 2006-03-01 13:54:26 +0100 (Wed, 01 Mar 2006) | 2 lines

 * Convert to table to reduce exe size.

........
r2720 | daniel | 2006-03-01 16:29:05 +0100 (Wed, 01 Mar 2006) | 3 lines

  - Remove usage of memory unit; things like the lowmemory function are
    flawed for the same reason memavail is flawed.

........
r2724 | jonas | 2006-03-03 20:29:07 +0100 (Fri, 03 Mar 2006) | 3 lines

  + added extra info about when parser_e_illegal_assignment_to_count_var
    can happen

........
r2746 | hajny | 2006-03-05 00:05:24 +0100 (Sun, 05 Mar 2006) | 1 line

  * Fix to allow / instead of " fexpand.inc
........
r2747 | hajny | 2006-03-05 00:09:06 +0100 (Sun, 05 Mar 2006) | 1 line

  * Slight optimization (constant instead of variable in assignment)
........
r2748 | tom_at_work | 2006-03-05 00:40:43 +0100 (Sun, 05 Mar 2006) | 2 lines

* changed library handle types to the correct TLibHandle
* avoid call of FreeLibrary() with zero, because it crashes on some linux versions
........
r2749 | tom_at_work | 2006-03-05 10:18:36 +0100 (Sun, 05 Mar 2006) | 1 line

* fixed opengl package compilation on Win32
........
r2750 | yury | 2006-03-05 10:48:31 +0100 (Sun, 05 Mar 2006) | 1 line

* fixed paszlib compiling in objfpc/delphi mode.
........
r2759 | marco | 2006-03-05 14:36:31 +0100 (Sun, 05 Mar 2006) | 3 lines

 * now first searches ppcross<target> if -P<target> is specified.
	After that it still searches for ppc<target>

........
r2761 | daniel | 2006-03-05 19:27:17 +0100 (Sun, 05 Mar 2006) | 2 lines

  * Add extra sequences for FreeBSD xterm

........
r2767 | peter | 2006-03-05 21:16:03 +0100 (Sun, 05 Mar 2006) | 2 lines

  * fixed xmlreg depends

........
r2768 | yury | 2006-03-05 21:36:16 +0100 (Sun, 05 Mar 2006) | 2 lines

* do not try to call cross compiler if host CPU was specified using -P parameter.
* tabs to spaces.
........
r2776 | peter | 2006-03-05 22:53:41 +0100 (Sun, 05 Mar 2006) | 2 lines

  * fixed invalid typecast

........
r2777 | daniel | 2006-03-05 23:00:20 +0100 (Sun, 05 Mar 2006) | 2 lines

  * Send escape code to xterm to make alt+key send an escape prefix.

........
r2787 | peter | 2006-03-06 12:25:09 +0100 (Mon, 06 Mar 2006) | 2 lines

  * fix for bug #4962, fix by Marton Pap

........

git-svn-id: branches/fixes_2_0@2797 -

+++ packages/base/paszlib/gzio.pas
@@ -46,7 +46,7 @@ function gzflush (f:gzFile; flush:integer)           : integer;
 function gzseek  (f:gzfile; offset:z_off_t; whence:integer) : z_off_t;
 function gztell  (f:gzfile) : z_off_t;
 function gzclose (f:gzFile)                      : integer;
-function gzerror (f:gzFile; var errnum:integer)      : string;
+function gzerror (f:gzFile; var errnum:smallint)      : string;
 function gzsetparams (f:gzfile; level:integer; strategy:integer) : integer;
 function gzrewind (f:gzFile) : integer;
 function gzeof (f:gzfile) : boolean;
@@ -1163,7 +1163,7 @@ end;
 
 ============================================================================}
 
-function gzerror (f:gzfile; var errnum:integer) : string;
+function gzerror (f:gzfile; var errnum:smallint) : string;
 var
  m : string;
  s : gz_streamp;
